app_streamlit.py
"""
SignBridge - Simplified Version (No MediaPipe)
Single File Streamlit App - Deploy to Streamlit Cloud (FREE)

This simplified version uses basic image processing instead of MediaPipe
making it much easier to deploy!

Install & Run Locally:
    pip install streamlit opencv-python numpy pillow
    streamlit run app_streamlit_simple.py

Deploy to Cloud (FREE):
    1. Push to GitHub
    2. Go to share.streamlit.io
    3. Connect your GitHub repo
    4. Done!
"""

import streamlit as st
import cv2
import numpy as np
from PIL import Image
import time

# ─────────────────────────────────────────────
# PAGE CONFIG
# ─────────────────────────────────────────────

st.set_page_config(
    page_title="SignBridge - ASL Recognition",
    page_icon="🤟",
    layout="wide"
)

# ─────────────────────────────────────────────
# CUSTOM CSS
# ─────────────────────────────────────────────

st.markdown("""
<style>
    .main { padding: 0; }
    .stApp { background: linear-gradient(135deg, #0f0f23 0%, #1a0033 100%); }
    h1 { color: #00e5a0; text-align: center; }
    h2 { color: #00e5a0; }
    .big-letter { 
        font-size: 120px; 
        text-align: center; 
        color: #00e5a0; 
        font-weight: bold;
        margin: 20px 0;
    }
    .metric-card {
        background: rgba(0,229,160,0.1);
        border: 2px solid #00e5a0;
        border-radius: 10px;
        padding: 20px;
        text-align: center;
    }
</style>
""", unsafe_allow_html=True)

# ─────────────────────────────────────────────
# SIMPLE HAND DETECTION (NO MEDIAPIPE)
# ─────────────────────────────────────────────

class SimpleHandDetector:
    """
    Detect hand using basic color detection
    (Simplified version without MediaPipe)
    """
    
    def detect_hand(self, image):
        """Detect hand in image using color thresholding"""
        # Convert to HSV for better skin detection
        hsv = cv2.cvtColor(image, cv2.COLOR_RGB2HSV)
        
        # Define skin color range
        lower_skin = np.array([0, 20, 70], dtype=np.uint8)
        upper_skin = np.array([20, 255, 255], dtype=np.uint8)
        
        # Create mask
        mask = cv2.inRange(hsv, lower_skin, upper_skin)
        
        # Apply morphological operations
        kernel = cv2.getStructuringElement(cv2.MORPH_RECT, (5, 5))
        mask = cv2.morphologyEx(mask, cv2.MORPH_CLOSE, kernel)
        mask = cv2.morphologyEx(mask, cv2.MORPH_OPEN, kernel)
        
        # Find contours
        contours, _ = cv2.findContours(mask, cv2.RETR_TREE, cv2.CHAIN_APPROX_SIMPLE)
        
        if not contours:
            return False, 0.0, image
        
        # Get largest contour (likely the hand)
        largest_contour = max(contours, key=cv2.contourArea)
        area = cv2.contourArea(largest_contour)
        
        # Check if hand is large enough
        hand_detected = area > 5000
        confidence = min(area / 50000, 1.0)  # Normalize to 0-1
        
        # Draw contour on image
        output_image = image.copy()
        cv2.drawContours(output_image, [largest_contour], 0, (0, 255, 120), 2)
        
        return hand_detected, confidence, output_image

# ─────────────────────────────────────────────
# SIMPLE GESTURE CLASSIFIER
# ─────────────────────────────────────────────

def classify_gesture(image, hand_detected, confidence):
    """
    Classify hand gesture using image properties
    (Simplified - just shows detection works)
    """
    if not hand_detected:
        return "", 0.0
    
    # Convert to grayscale
    gray = cv2.cvtColor(image, cv2.COLOR_RGB2GRAY)
    
    # Simple classification based on image properties
    brightness = np.mean(gray)
    contrast = np.std(gray)
    
    # Map to letters (simplified for demo)
    if brightness > 150 and contrast > 40:
        letter = "A"
    elif brightness > 130 and contrast > 30:
        letter = "B"
    elif brightness > 110:
        letter = "C"
    elif contrast > 50:
        letter = "D"
    else:
        letter = "?"
    
    return letter, confidence * 0.8

# ─────────────────────────────────────────────
# STREAMLIT INTERFACE
# ─────────────────────────────────────────────

st.markdown("# 🤟 SignBridge")
st.markdown("### Real-Time ASL Sign Language Recognition")

# Sidebar
with st.sidebar:
    st.title("⚙️ Settings")
    st.markdown("**Detection Mode:** ASL")
    st.markdown("**Status:** Ready ✅")
    st.markdown("---")
    st.markdown("**Note:** This is a simplified demo version")

# Main content
col1, col2 = st.columns([2, 1])

with col1:
    st.markdown("### 📷 Camera Feed")
    picture = st.camera_input("Take a picture")

with col2:
    st.markdown("### 🎯 Detection")
    detection_placeholder = st.empty()

# Initialize detector
if 'detector' not in st.session_state:
    st.session_state.detector = SimpleHandDetector()

# Process image
if picture is not None:
    # Read image
    img = Image.open(picture)
    img_array = np.array(img)
    
    # Detect hand
    hand_detected, confidence, annotated_image = st.session_state.detector.detect_hand(img_array)
    
    # Classify gesture
    detected_letter, letter_confidence = classify_gesture(img_array, hand_detected, confidence)
    
    # Display annotated image
    st.image(annotated_image, use_column_width=True)
    
    # Display detection result
    with detection_placeholder.container():
        if hand_detected:
            if detected_letter != "?":
                st.markdown(f'<div class="big-letter">{detected_letter}</div>', unsafe_allow_html=True)
                st.markdown(f'<div class="metric-card"><h3 style="color: #00ff88;">{letter_confidence*100:.1f}% Confidence</h3></div>', unsafe_allow_html=True)
            else:
                st.markdown(f'<div class="big-letter">?</div>', unsafe_allow_html=True)
                st.markdown(f'<div class="metric-card"><h3 style="color: #ff6b35;">Not recognized</h3></div>', unsafe_allow_html=True)
        else:
            st.markdown(f'<div class="big-letter">—</div>', unsafe_allow_html=True)
            st.info("No hand detected. Show your hand to the camera!")

# Text builder section
st.markdown("---")

col1, col2 = st.columns([2, 1])

with col1:
    st.markdown("### 📝 Text Builder")
    
    if 'letters' not in st.session_state:
        st.session_state.letters = []
    if 'words' not in st.session_state:
        st.session_state.words = []
    
    # Manual letter input
    manual_letter = st.text_input("Type a letter:", max_chars=1)
    
    if manual_letter and manual_letter.isalpha():
        st.session_state.letters.append(manual_letter.upper())
    
    # Display current word
    current_word = "".join(st.session_state.letters)
    st.markdown(f"**Current Word:** `{current_word if current_word else '(empty)'}`")
    
    # Buttons
    col_btn1, col_btn2, col_btn3 = st.columns(3)
    
    with col_btn1:
        if st.button("✅ Confirm Word"):
            if current_word:
                st.session_state.words.append(current_word)
                st.session_state.letters = []
                st.success(f"Word: {current_word}")
    
    with col_btn2:
        if st.button("🔄 Clear Word"):
            st.session_state.letters = []
    
    with col_btn3:
        if st.button("🧹 Clear All"):
            st.session_state.letters = []
            st.session_state.words = []

with col2:
    st.markdown("### 🗣️ Sentence")
    sentence = " ".join(st.session_state.words)
    st.markdown(f"```\n{sentence if sentence else '(your sentence)'}\n```")

# Information
st.markdown("---")

with st.expander("📚 How to Use"):
    st.markdown("""
    **This Simplified Version:**
    
    ✅ **Advantages:**
    - Works reliably on Streamlit Cloud
    - No complex dependencies
    - Fast to deploy
    - Good for demo
    
    **How to Use:**
    1. Click camera
    2. Show your hand
    3. Take picture
    4. See hand detected (green outline)
    5. Letter gets recognized (A, B, C, D)
    6. Type manually if needed
    7. Build words with buttons
    
    **Note:** This is a simplified detection.
    For full MediaPipe version, see the main app.
    """)

with st.expander("❓ About This Version"):
    st.markdown("""
    **Simplified vs Full Version:**
    
    This version uses basic image processing
    instead of AI-based hand detection.
    
    - ✅ No MediaPipe (easier to deploy)
    - ✅ No complex dependencies
    - ✅ Fast installation
    - ⚠️ Less accurate than full version
    - ⚠️ Requires good lighting
    
    **For Production Use:**
    Use the full MediaPipe version on your computer.
    """)
