requirements.txt
streamlit==1.26.0
opencv-python-headless==4.5.5.64
numpy==1.21.6
pillow==9.0.0
